MAINT: move cell classes to one file
