+++ hnn_core/__init__.py
@@ -3,8 +3,7 @@ from .drives import drive_event_times
 from .params import Params, read_params
 from .network import Network
 from .cell_response import CellResponse, read_spikes
-from .pyramidal import L2Pyr, L5Pyr
-from .basket import L2Basket, L5Basket
+from .cell_classes import L2Pyr, L5Pyr, L2Basket, L5Basket
 from .parallel_backends import MPIBackend, JoblibBackend
 
 __version__ = '0.2.dev0'
